% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{clip_raster}
\alias{clip_raster}
\title{Clip raster}
\usage{
clip_raster(x, y)
}
\arguments{
\item{x}{raster to clip}

\item{y}{raster with clipping extent}
}
\description{
Clips a raster with another raster by cropping, then multiplying their values
}
