% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_vegpar.R
\name{get_basin_avg}
\alias{get_basin_avg}
\title{Get Basin Average}
\usage{
get_basin_avg(r)
}
\arguments{
\item{r}{Input raster stack of monthly average parameter values}
}
\description{
Calculates the basin average values of a distributed parameter, given monthly maps
}
\details{
This function is not limited to vegetation parameters. t's just a function for calculating the average value of a raster (stack).
}
\examples{
LAI_wt_avg <- vegpar_wt_avg(LAI, CV)
spatial_avg_LAI <- get_basin_avg(LAI_wt_avg)
}
