% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{lineplot2}
\alias{lineplot2}
\title{Line Plot 2}
\usage{
lineplot2(x, y1, y2, legtext, lpos = "topleft", sz = 1.5, lw = 1.5, ...)
}
\arguments{
\item{x}{x-coordinates}

\item{y1}{y-coordinates for second data series}

\item{...}{additional arguments for plot()}
}
\description{
Makes a line plot comparing two data series
}
