% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{majority_raster}
\alias{majority_raster}
\title{Majority Raster}
\usage{
majority_raster(cv, thres)
}
\arguments{
\item{cv}{raster stack with land cover fraction data}

\item{thres}{threshold for determining whether a pixel is mostly that land cover}
}
\description{
Calculate which grid cells are mostly each kind of LC cover
}
\examples{
CVmaj <- majority_raster(cv_clip, 0.9)
}
