% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{make_spatial_polygon}
\alias{make_spatial_polygon}
\title{Make Spatial Polygon}
\usage{
make_spatial_polygon(e)
}
\arguments{
\item{e}{extent \link{xmin, xmax, ymin, ymax}}
}
\description{
Makes a rectangular spatial polygon from four points
}
