% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf2raster.R
\name{netcdf2raster}
\alias{netcdf2raster}
\title{NetCDF to raster}
\usage{
netcdf2raster(ncdf_name, varname, classname, outdir)
}
\arguments{
\item{ncdf_name}{netcdf filename}

\item{varname}{name of NetCDF variable (e.g. LAI or fcanopy)}

\item{classname}{list of vegetation class names}

\item{outdir}{directory to save outputs}
}
\value{
raster containing vegetation parameter data from NetCDF file
}
\description{
Converts a 3D or 4D NetCDF file to a raster. Saves output as geotiff. Specifically designed for vegetation parameters from the VIC-5 (Image Mode) parameter file.
}
\examples{
vegpar_name <- "params.CONUS_MX.MOD_IGBP.mode.2000_2016.nc"
class.list <- "igbp_classnames.txt"
classname <- read.table(class.list, stringsAsFactors = FALSE)[[1]]
geotiff.dir <- "/home/jschap/Documents/"
netcdf2raster(vegpar_name, "LAI", classname, geotiff.dir)
# class.list is a text file containing the names of each vegetation class. The order might be important.
#' @details Assumes NetCDF file is set up as [x, y, time, (class)]
}
