% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_vegpar.R
\name{stack_by_time}
\alias{stack_by_time}
\title{Stack by Time}
\usage{
stack_by_time(varname, outdir, nt = 12)
}
\arguments{
\item{varname}{name of variable (e.g. LAI)}

\item{outdir}{directory to save outputs}

\item{nt}{number of time steps. Default is 12 (monthly)}
}
\value{
par.bytime, a list of raster stacks.
}
\description{
Rearranges the vegetation parameter files into a list of raster stacks for each time t
}
\details{
Returns a list of raster stacks, where each entry is for a different time,
and each layer contains vegetation parameter data for a different land cover.
This code is set up for monthly average data.
}
