% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_vegpar.R
\name{subset_vegpar}
\alias{subset_vegpar}
\title{Subset vegpar}
\usage{
subset_vegpar(indir, varname, classname, outdir, basin_mask, nt = 12)
}
\arguments{
\item{indir}{directory containing geotiff files with vegetation parameter data}

\item{varname}{name of vegetation parameter (e.g. LAI or fcanopy)}

\item{classname}{list of vegetation class names}

\item{outdir}{directory to save outputs}

\item{basin_mask}{basin mask}

\item{nt}{number of time steps (generally going be to 12 (months))}
}
\value{
the most recent clipped raster
}
\description{
Subsets a vegetation parameter to a basin mask
}
\details{
Works specifically with the UMD-NLDAS land cover classes or the IGBP land cover classes.
To do: Make this more general by calling this code for a particular filename, instead of for a group of files at once.
}
\examples{
ucrb_mask <- raster("colo_mask.tif")
class.list <- "igbp_classnames.txt"
classname <- read.table(class.list, stringsAsFactors = FALSE)[[1]]
geotiff.dir <- "~/Documents"
clipped.gtiff.dir <- "~/Documents/clipped_to_ucrb"
subset_vegpar(geotiff.dir, "LAI", classname, clipped.gtiff.dir, ucrb_mask)
}
