% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_vegpar.R
\name{vegpar_seasonal_avg}
\alias{vegpar_seasonal_avg}
\title{Vegpar Seasonal Average}
\usage{
vegpar_seasonal_avg(meanrasters)
}
\arguments{
\item{meanrasters}{Input raster stack of monthly average vegpar values}
}
\value{
raster brick containing seasonal and annual average vegetation parameter values
}
\description{
Calculates the seasonal value of the vegetation parameter, given monthly average maps
}
\details{
Assumes the inputs are 1. Use output fromvegpar_wt_avg().
}
\examples{
LAI <- stack_by_time("LAI", clipped.gtiff.dir)
CV <- stack(cvnames)
LAI_wt_avg <- vegpar_wt_avg(LAI, CV)
LAI_seasonal_avg <- vegpar_seasonal_average(LAI)
}
