% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_vegpar.R
\name{vegpar_wt_avg}
\alias{vegpar_wt_avg}
\title{Vegpar Weighted Average}
\usage{
vegpar_wt_avg(vpar, CV)
}
\arguments{
\item{vpar}{Input list of LAI raster stacks (LAI, fcanopy, albedo, etc.).}

\item{CV}{Input raster stack with cover fractions}
}
\value{
raster stack containing weighted average vegetation parameter values
}
\description{
Calculates the average value of the vegetation parameter in each month, taking into account land-cover based weighting.
}
\details{
Generate the input list vpar with stack_by_time()
}
